requirements.txt
-e .
typer
requests
dotenv
pyyaml
